feat: abstract away Wazero's api.ValueType and api.EncodeXX/api.DecodeXX (#44)
Currently we are exposing these wazero APIs:
-
[`RuntimeConfig`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/config.go#L40-L171)
-
[`ModuleConfig`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/config.go#L444-L638)
-
[`api.ValueType`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/api/wasm.go#L83-L110)
-
[`api.EncodeXX/api.DecodeXX`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/api/wasm.go#L693-L755)
-
[`api.Memory`](https://github.com/tetratelabs/wazero/blob/d39d84505ad8021b5ba47bbc9dd55911b9b0406f/api/wasm.go#L557-L673)

The idea of this PR is hide away wazero from normal usage. If the user
doesn't need any advanced configuration or functionality, then they
shouldn't have to know about Wazero. That's why we are abstracting away
`api.EncodeXX/api.DecodeXX` and `api.ValueType` since they will be used
whenever you write a host function.

`RuntimeConfig` and `ModuleConfig` are about giving the user fine grain
control of the underyling configurations. We expose some of the
functionality as part of the manifest as well, [see this for a more
detailed
breakdown](https://github.com/extism/go-sdk/pull/43#issuecomment-1833735502).

And `api.Memory` is for write-through access of the underyling memory.

The only thing I am a bit worried about is
[`RuntimeConfig.WithCloseOnContextDone`](https://github.com/tetratelabs/wazero/blob/5796897f37852bf2042cd63959ce4ff673c4366e/config.go#L170C2-L170C24)
because by default wazero sets it to false, which might be very
confusing. We do set it to true if the user configures a timeout though.