Correctly use context in plugin and provide alternative _WithContext methods
This commit ensures that the extism plugin struct does
not store a context object in itself as this is not idiomatic
Go and leads to unexpected behavior. Instead, a context object
is passed to each plugin method call that runs a long running
operation.

To not break the interface of the SDK, the existing SDK method
signatures were preserved, but they were changed to use
context.Background under the hood. Next to these, alternative
_WithContext methods were added, which allow the caller to
provide a custom context object.