feat!: provide plugins get_log_level runtime function & support levels (#74)
This PR aligns the extism host env with Rust, as per
https://github.com/extism/extism/pull/758

NOTE of breaking change:

I've removed the ability to set the log level on the `*Plugin` itself.
To align with Rust SDK, this is only an SDK-global setting. You can
still add a custom logger to each plugin function. The SDK won't print
any logs unless there is a log function set for each plugin. The global
setting is in line with how libraries are expected to respect the main
application's log level.