Support concurrent plugin execution through instances (#81)
See [Discord
thread](https://discord.com/channels/1011124058408112148/1050087851443888138/1298037616838443029).

It seems like there should be a `Plugin` type that is thread-safe, and a
`PluginInstance` type that is not thread-safe. The `Plugin `contains the
runtime and the compiled module, and you'd have something like `func(p
*Plugin) NewInstance() *PluginInstance` which would instantiate a wazero
module that would then be used in a single-threaded context.

The idea here is I can compile my plugin one time, and call functions
from the module as many times as I want, concurrently, without having to
re-compile.